Changing double to decimal data type

+++ MAM_Assessment/Book.cs
@@ -3,7 +3,7 @@ namespace MAM_Assessment
 {
     public class Book : Product
     {
-        public Book(double price, string description, int quantity) : base(price, description, quantity)
+        public Book(decimal price, string description, int quantity) : base(price, description, quantity)
         {
 
         }
