+++ MAM_Assessment/Cosmetics.cs
@@ -3,7 +3,7 @@ namespace MAM_Assessment
 {
     public class Cosmetics : Product
     {
-        public Cosmetics(double price, string description, int quantity) : base(price, description, quantity)
+        public Cosmetics(decimal price, string description, int quantity) : base(price, description, quantity)
         {
 
         }
