+++ MAM_Assessment/IProduct.cs
@@ -5,9 +5,9 @@ namespace MAM_Assessment
 {
     public interface IProduct
     {
-        double CalculateTotalTax(double price);
-        double CalculateTotalImportTax(double price);
-        double CalculateImportTax(double price);
+        decimal CalculateTotalTax(decimal price);
+        decimal CalculateTotalImportTax(decimal price);
+        decimal CalculateImportTax(decimal price);
         void DisplayProducts(List<Product> products);
     }
 }
